modify abilene topo

+++ experiments/testbed/configure-kulfi.sh
@@ -2,14 +2,18 @@
 . utils/colors.sh
 . utils/common.sh
 cp $KULFI_GIT_DIR/agent/agent.py $ATLAS_KULFI_CONFIG/
+KMUPDATE=$1
 
-echo ${CYAN}------ Copy Kernel Module -----${RESTORE}
-rm -rf $ATLAS_KULFI_CONFIG/kernel
+if ! [[ -z $KMUPDATE ]]; then
+	echo ${CYAN}------ Copy Kernel Module -----${RESTORE}
+	rm -rf $ATLAS_KULFI_CONFIG/kernel
+	cp -r $KULFI_GIT_DIR/kernel/ $ATLAS_KULFI_CONFIG/
+fi
+
 rm -rf $ATLAS_KULFI_CONFIG/routes
-cp -r $KULFI_GIT_DIR/kernel/ $ATLAS_KULFI_CONFIG/
 cp -r $KULFI_GIT_DIR/routes/ $ATLAS_KULFI_CONFIG/
 
-dsh -M -g atlas-abilene -c "rm -rf $ATLAS_KULFI_CONFIG ; scp -r olympic:$ATLAS_KULFI_CONFIG ./ ; $ATLAS_KULFI_CONFIG/configure.sh"
+dsh -M -g atlas-abilene -c "rm -rf $ATLAS_KULFI_CONFIG ; scp -r olympic:$ATLAS_KULFI_CONFIG ./ ; $ATLAS_KULFI_CONFIG/configure.sh $KMUPDATE"
 sleep 1
 # Verify modkulfi is loaded on all servers
 MOD=`dsh -M -g atlas-abilene -c "sudo lsmod | grep kulfi " | wc -l`
